% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dt}
\alias{example_dt}
\title{Synthetic lethal RNAi screen example data.}
\format{
A data.table with 4320 rows and 8 variables
}
\usage{
data(example_dt)
}
\value{
A data.table containing RANi screen data, the READOUT value has no real biological meaning.
}
\description{
A dataset containing synthetic lethal RNAi screen data to show how functions work.
The variables are as follows (all are character except READOUT):
}
\details{
\itemize{
  \item PLATE. plate names.
  \item MASTER_PLATE. master plate names.
  \item WELL_CONTENT_NAME. siRNA targets of wells.
  \item EXPERIMENT_TYPE. sample, negative/positive controls.
  \item EXPERIMENT_MODIFICATION. experiment conditions, "treatment" or "control".
  \item ROW_NAME. row names of plates.
  \item COL_NAME. column names of plates.
  \item READOUT. screen results.
}
}
\keyword{datasets}
