% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exampleDat.R
\docType{data}
\name{exampleDat}
\alias{exampleDat}
\title{Synthetic lethal RNAi screen example data.}
\format{A data frame with 4320 rows and 8 variables}
\usage{
data(exampleDat)
}
\value{
A data frame containing RANi screen data, the READOUT value has no real biological meaning.
}
\description{
A dataset containing synthetic lethal RNAi screen data to show how functions work.
The variables are as follows:
}
\details{
\itemize{
  \item PLATE. plate names.
  \item MASTER_PLATE. master plate names.
  \item WELL_CONTENT_NAME. siRNA targets of wells.
  \item EXPERIMENT_TYPE. sample, negative/positive controls.
  \item EXPERIMENT_MODIFICATION. experiment conditions, "treatment" or "control".
  \item ROW_NAME. row names of plates.
  \item COL_NAME. column names of plates.
  \item READOUT. screen results.
}
}
\keyword{datasets}

