% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot of RNAi screen results}
\usage{
scatterPlot(
  dta,
  scatter_colour = rainbow(10),
  controlOnly = FALSE,
  control_name = NULL
)
}
\arguments{
\item{dta}{synthetic lethal RNAi screen data}

\item{scatter_colour}{colour for different signals}

\item{controlOnly}{whether or not to plot control wells only}

\item{control_name}{names of control siRNAs.}
}
\value{
a ggplot object
}
\description{
Produce a single plot for readous of each plate, with the option of highlighting
specific signals, like positive/negative controls.
}
\examples{
data(example_dt)
scatterPlot(example_dt, controlOnly = FALSE,
           control_name = c("PLK1 si1", "scrambled control si1", "lipid only"))
}
