% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTest.R
\name{tTest}
\alias{tTest}
\title{student's t-test on B-score}
\usage{
tTest(mtx, bScore, n_treat, n_cont)
}
\arguments{
\item{bScore}{normalized bScore}

\item{n_treat}{number of treatment plates}

\item{n_cont}{number of control plates}

\item{masterPlate}{the master plate to be analyzed}
}
\value{
A list containing student's t-test for each master plate
\itemize{
  \item pvalue: p-value of the t-test
  \item Treat_Cont: difference in bscore: treatment - control
  \item p_adj: BH adjusted p-value
}
}
\description{
Select hits by student's t-test using B-score from treatment and control plates.
}
\examples{
bscore_res <- sapply(unique(exampleDat$MASTER_PLATE), bScore,
  exampleDat, control = "control", treatment = "treatment", simplify = FALSE)
bscore_ttest  <- sapply(names(bscore_res), tTest, bscore_res, n_treat = 3,
  n_cont = 3, simplify = FALSE, USE.NAMES = TRUE)
bscore_comb <- data.frame(do.call(rbind, lapply(names(bscore.ttest),
  function(x) if (!is.null(bscore.ttest[[x]])) {data.frame(MASTER_PLATE = x,
  siRNAs = rownames(bscore.ttest[[x]]), bscore.ttest[[x]])})))
}
\references{
Birmingham, A. et al. Statistical methods for analysis of high-throughput RNA interference screens. Nat Methods 6, 569-575 (2009).
}
