% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateHeatmap.R
\name{plateHeatmap}
\alias{plateHeatmap}
\title{Heatmap of all plates}
\usage{
plateHeatmap(dta, base_size = 12, heatmap_col = NULL)
}
\arguments{
\item{dta}{synthetic lethal RNAi screen data}

\item{base_size}{basic font size used for x/y axis and title for heatmaps}

\item{heatmap_col}{color function generated by colorRampPalette.}
}
\value{
a ggplot object
}
\description{
Put all individual plates in one graph, values are the readout in experiments.
}
\examples{
data(example_dt)
plateHeatmap(example_dt)
}
