% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plateHeatmap.R
\name{plateHeatmap}
\alias{plateHeatmap}
\title{Heatmap of all plates}
\usage{
plateHeatmap(dat, baseSize = 12)
}
\arguments{
\item{dat}{synthetic lethal RNAi screen data}

\item{baseSize}{basic font size used for x/y axis and title for heatmaps}
}
\value{
a ggplot object
}
\description{
Put all individual plates in one graph, values are the readout in experiments.
}
\examples{
tem.1 <- plateHeatmap(exampleDat)
ggsave("platesHeatmap.pdf", plot = tem.1, width = 500, height = 500, limitsize = FALSE)
}

