% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Scatter plot of RNAi screen results}
\usage{
scatterPlot(dat, controlOnly = FALSE, colour, ...)
}
\arguments{
\item{dat}{synthetic lethal RNAi screen data}

\item{controlOnly}{whether or not to plot control wells only}

\item{colour}{colour for different signals}

\item{...}{positive/negative signals, must be specified}
}
\value{
a ggplot object
}
\description{
Produce a single plot for readous of each plate, with the option of highlighting
specific signals, like positive/negative controls.
}
\examples{
scatterPlot(exampleDat, controlOnly = FALSE, colour = rainbow(10),
  "PLK1 si1", "scrambled control si1", "lipid only")
}

