% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siRNAPlot.R
\name{siRNAPlot}
\alias{siRNAPlot}
\title{Plot siRNA data and quality metrics.}
\usage{
siRNAPlot(gene, dat, controlsiRNA, FILEPATH = ".", colour = rainbow(10),
  zPrimeMed, zPrimeMean, treatment, control, normMethod = c("PLATE"),
  width = 15, height = 14)
}
\arguments{
\item{gene}{gene symbol, case sensitive}

\item{dat}{synthetic lethal RNAi screen data}

\item{controlsiRNA}{controlsiRNA could be a vector of several siRNA, including postive/negative control}

\item{FILEPATH}{path to store the figure}

\item{colour}{colour used in graphs}

\item{zPrimeMed}{zPrime factor basing on median}

\item{zPrimeMean,}{zPrime factor basing on mean}

\item{treatment}{the treatment condition in EXPERIMENT_MODIFICATION}

\item{control}{the control condition in EXPERIMENT_MODIFICATION}

\item{normMethod}{could be a PLATE and negative controls}

\item{width}{width of the plot}

\item{height}{height of the plot}
}
\value{
Return the ggplot2 objects in a list, which could be plotted individually.
}
\description{
Plot the normalized RNAi screen data, row data, control signals and Z' factor.
}
\examples{
zF_mean <- zFactor(exampleDat, negativeCon = "scrambled control si1",
  positiveCon = "PLK1 si1")
zF_med <- zFactor(exampleDat, negativeCon = "scrambled control si1",
  positiveCon = "PLK1 si1", useMean = FALSE)
tem.1 <- siRNAPlot("AAK1", exampleDat,
  controlsiRNA = c("lipid only", "scrambled control si1"),
  FILEPATH = ".",  zPrimeMed = zF_med, zPrimeMean = zF_mean,
  treatment = "treatment", control = "control",
  normMethod = c("PLATE", "lipid only", "scrambled control si1"))
}

