% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tTest.R
\name{tTest}
\alias{tTest}
\title{student's t-test basing on B-score}
\usage{
tTest(masterPlate, bScore, numTreat, numCont)
}
\arguments{
\item{masterPlate}{the master plate to be analyzed}

\item{bScore}{normalized bScore}

\item{numTreat}{number of treatment plates}

\item{numCont}{number of control plates}
}
\value{
A list containing student's t-test for each master plate
\itemize{
  \item pvalue: p-value of the t-test
  \item Treat_Cont: difference in bscore: treatment - control
  \item p_adj: BH adjusted p-value
}
}
\description{
Select hits by student's t-test using B-score from treatment and control plates.
}
\examples{
bscore.res <- sapply(as.character(unique(exampleDat$MASTER_PLATE)), bScore,
  exampleDat, control = "control", treatment = "treatment", simplify = FALSE)
bscore.ttest  <- sapply(names(bscore.res), tTest, bscore.res, numTreat = 3,
  numCont = 3, simplify = FALSE, USE.NAMES = TRUE)
bscore.combined <- data.frame(do.call(rbind, lapply(names(bscore.ttest),
  function(x) if (!is.null(bscore.ttest[[x]])) {data.frame(MASTER_PLATE = x,
  siRNAs = rownames(bscore.ttest[[x]]), bscore.ttest[[x]])})))
}
\references{
Birmingham, A. et al. Statistical methods for analysis of high-throughput RNA interference screens. Nat Methods 6, 569-575 (2009).
}

